+++ src/main/java/racingcar/controller/RacingGame.java
@@ -10,6 +10,7 @@ import static racingcar.view.ConsoleOutput.printNewLine;
 import static racingcar.view.ConsoleOutput.printProgressState;
 
 import java.util.List;
+import racingcar.domain.CarManager;
 import racingcar.domain.FinalResult;
 import racingcar.domain.ProgressState;
